Address agent graph review feedback
