+++ codex-rs/agent-graph-store/src/lib.rs
@@ -1,14 +1,12 @@
 //! Storage-neutral parent/child topology for thread-spawned agents.
 
 mod error;
-mod in_memory;
 mod local;
 mod store;
 mod types;
 
 pub use error::AgentGraphStoreError;
 pub use error::AgentGraphStoreResult;
-pub use in_memory::InMemoryAgentGraphStore;
 pub use local::LocalAgentGraphStore;
 pub use store::AgentGraphStore;
 pub use store::AgentGraphStoreFuture;
